+++ dev_utils/views.py
@@ -1,4 +1,5 @@
 from django.views.generic.base import TemplateView
+from django.contrib import messages
 
 
 # Create your views here.
@@ -79,7 +80,16 @@ class MultipleFormView(TemplateView):
             return self.process_forms(forms_initialized)
         else:
             context = merge_dicts(self.get_context_data(), forms_initialized)
+            context = self.post_valid_reject(context)
+            messages.add_message(
+                self.request,
+                messages.ERROR,
+                "Something is wrong. Make sure you're filling out the form properly.",
+            )
             return self.render_to_response(context)
 
     def process_forms(self, form_instances):
+        raise NotImplementedError
+
+    def post_valid_reject(context):
         raise NotImplementedError